- Changed Message into an interface, default impl is BaseMessage (https://issues.jboss.org/browse/JGRP-2218) - Added impls BytesMessage (default) and ObjectMessage, test suite passes 100% - Changed FRAG, FRAG2, FRAG3, COMPRESS, SYM/ASYM_ENCRYPT to fragment messages other than BytesMessages (e.g. ObjectMessages) - Changed Message.size() to return int instead of long - Moved Message.setTransientFlag(TransientFlag... flags) to setFlag(TransientFlag... flags) - Renamed Buffer -> ByteArray - SEQUENCER now handles messages other than BytesMessage correctly - Added ObjectMessageTest - StreamableTo/FromBuffer() and messageTo/FromBuffer now throw an exception - Created EmptyMessage and using it instead of ByteMessage to send messages that contain only headers but no payload - Added ObjectMessage / ObjectMessageSerializable - New class PartialDataOutputStream - Added FragmentedMessage - Added FRAG4 (using FragmentedMessage) - Added FRAG4_Test

+++ src/org/jgroups/protocols/BaseBundler.java
@@ -54,8 +54,8 @@ public abstract class BaseBundler implements Bundler {
     public int size() {
         lock.lock();
         try {
-            long num=msgs.values().stream().flatMap(Collection::stream).map(Message::size).reduce(0L, (a, b) -> a + b);
-            return (int)num;
+            int num=msgs.values().stream().flatMap(Collection::stream).map(Message::size).reduce(0, (a, b) -> a + b);
+            return num;
         }
         finally {
             lock.unlock();
@@ -124,7 +124,7 @@ public abstract class BaseBundler implements Bundler {
         }
     }
 
-    @GuardedBy("lock") protected void addMessage(Message msg, long size) {
+    @GuardedBy("lock") protected void addMessage(Message msg, int size) {
         Address dest=msg.getDest();
         List<Message> tmp=msgs.computeIfAbsent(dest, k -> new ArrayList<>(5));
         tmp.add(msg);
